requirements.txt
#
# Python dependencies for IT Service Request Tracking System
#
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0